Refactor trace writer detail state helpers
