+++ crates/rulemorph_trace/src/trace_writer/detail.rs
@@ -1,5 +1,7 @@
 use serde_json::Value as JsonValue;
 
+use super::options::TraceDetailLevel;
+
 pub(super) fn append_detail_reason(existing: Option<String>, reason: &str) -> Option<String> {
     match existing {
         None => Some(reason.to_string()),
@@ -13,6 +15,30 @@ pub(super) fn append_detail_reason(existing: Option<String>, reason: &str) -> Op
                 Some(format!("{current},{reason}"))
             }
         }
+    }
+}
+
+pub(super) fn initial_detail_reasons(reason: Option<&str>) -> Vec<String> {
+    let mut detail_reason = Vec::new();
+    if let Some(reason) = reason {
+        for item in reason.split(|ch| ch == ',' || ch == ';') {
+            let trimmed = item.trim();
+            if trimmed.is_empty() {
+                continue;
+            }
+            if !detail_reason.iter().any(|existing| existing == trimmed) {
+                detail_reason.push(trimmed.to_string());
+            }
+        }
+    }
+    detail_reason
+}
+
+pub(super) fn detail_status_for_level(detail_level: TraceDetailLevel) -> String {
+    match detail_level {
+        TraceDetailLevel::Full => "full".to_string(),
+        TraceDetailLevel::Basic => "basic".to_string(),
+        TraceDetailLevel::Off => "dropped".to_string(),
     }
 }
